Many improvements

+++ Buttons/Upgrade.py
@@ -10,20 +10,24 @@ from helpers.ShipHelper import PlayerShip
 class UpgradeButtons:
 
     @staticmethod
-    async def startUpgrade(game: GamestateHelper, player, interaction: discord.Interaction, button:bool):
+    async def startUpgrade(game: GamestateHelper, player, interaction: discord.Interaction, button:bool, discTileUpgrade:str):
         ships = ["interceptor","cruiser","dread","starbase"]
         drawing = DrawHelper(game.gamestate)
         image = drawing.player_area(player)
         view = View()
         actions = str(player['upgrade_apt'])
         for ship in ships:
-            view.add_item(Button(label=ship.capitalize(), style=discord.ButtonStyle.blurple, custom_id=f"FCID{player['color']}_upgradeShip_{actions}_{ship}"))
+            view.add_item(Button(label=ship.capitalize(), style=discord.ButtonStyle.blurple, custom_id=f"FCID{player['color']}_upgradeShip_{actions}_{ship}_{discTileUpgrade}"))
 
-        if button:
-            view.add_item(Button(label="End Turn", style=discord.ButtonStyle.red, custom_id="endTurn"))
+        if button and discTileUpgrade != "dummy":
+            view.add_item(Button(label="End Turn", style=discord.ButtonStyle.red, custom_id=f"FCID{player['color']}_endTurn"))
             view.add_item(Button(label="Restart Turn", style=discord.ButtonStyle.gray, custom_id=f"FCID{player['color']}_restartTurn"))
             await interaction.message.delete()
-        await interaction.response.send_message(file=drawing.show_player_ship_area(image),ephemeral=True)
+        if discTileUpgrade == "dummy":
+            await interaction.response.send_message(file=drawing.show_player_ship_area(image),ephemeral=True)
+        else:
+            await interaction.followup.send(file=drawing.show_player_ship_area(image),ephemeral=True)
+            view.add_item(Button(label="Save For Future Upgrade Action", style=discord.ButtonStyle.red, custom_id=f"FCID{player['color']}_deleteMsg"))
         await interaction.followup.send(
             f"{interaction.user.mention}, choose which ship you would like to upgrade.", view=view)
 
@@ -32,53 +36,65 @@ class UpgradeButtons:
         view = View()
         actions = customID.split("_")[1]
         ship = customID.split("_")[2]
+        discTileUpgrade = customID.split("_")[3]
         player_helper.setOldShipParts(ship)
         game.update_player(player_helper)
         with open("data/parts.json", "r") as f:
             part_stats = json.load(f)
         for i in set(player[f"{ship}_parts"]):
-            view.add_item(Button(label=part_stats[i]["name"], style=discord.ButtonStyle.blurple, custom_id=f"FCID{player['color']}_selectOldPart_{actions}_{ship}_{i}"))
+            view.add_item(Button(label=part_stats[i]["name"], style=discord.ButtonStyle.blurple, custom_id=f"FCID{player['color']}_selectOldPart_{actions}_{ship}_{i}_{discTileUpgrade}"))
         await interaction.response.edit_message(content=f"{interaction.user.mention}, pick which part to replace or remove.",
                                                 view=view)
     @staticmethod
-    async def selectOldPart(game: GamestateHelper, player, interaction: discord.Interaction, customID : str):
+    async def selectOldPart(game: GamestateHelper, player, interaction: discord.Interaction, customID : str, player_helper:PlayerHelper):
         view = View()
         actions = customID.split("_")[1]
         ship = customID.split("_")[2]
         oldPart = customID.split("_")[3]
-        available_techs = ["empty", "ioc", "elc", "nud", "hul", "gas", "nus"]
+        discTileUpgrade = customID.split("_")[4]
+
+        if discTileUpgrade != "dummy":
+            await UpgradeButtons.chooseUpgrade(game, player, interaction, customID+"_"+discTileUpgrade,player_helper)
+            return
+
+
+        available_parts = ["ioc", "elc", "nud", "hul", "gas", "nus","empty"]
+        drawing = DrawHelper(game.gamestate)
         with open("data/parts.json", "r") as f:
             part_stats = json.load(f)
         for tech in part_stats:
             if tech in player["military_tech"]:
-                available_techs.append(tech)
+                available_parts.append(tech)
             if tech in player["grid_tech"]:
-                available_techs.append(tech)
+                available_parts.append(tech)
             if tech in player["nano_tech"]:
-                available_techs.append(tech)
+                available_parts.append(tech)
             if "ancient_parts" in player and tech in player["ancient_parts"]:
-                available_techs.append(tech)
-        for i in available_techs:
-            view.add_item(Button(label=part_stats[i]["name"], style=discord.ButtonStyle.blurple, custom_id=f"FCID{player['color']}_chooseUpgrade_{actions}_{ship}_{oldPart}_{i}"))
+                available_parts.append(tech)
+        available_parts = set(available_parts)
+        for i in available_parts:
+            view.add_item(Button(label=part_stats[i]["name"], style=discord.ButtonStyle.blurple, custom_id=f"FCID{player['color']}_chooseUpgrade_{actions}_{ship}_{oldPart}_{i}_{discTileUpgrade}"))
         await interaction.response.edit_message(content=f"{interaction.user.mention}, replace "
                                                         f"{part_stats[oldPart]['name']} with which part? Remove as a free action by selecting 'Empty'.", view=view)
+        await interaction.followup.send("Available parts", file=drawing.availablePartsFile(available_parts),ephemeral=True)
     @staticmethod
     async def chooseUpgrade(game: GamestateHelper, player, interaction: discord.Interaction, customID : str,player_helper : PlayerHelper):
         actions = int(customID.split("_")[1])
         ship = customID.split("_")[2]
         oldPart = customID.split("_")[3]
         newPart = customID.split("_")[4]
+        discTileUpgrade = customID.split("_")[5]
         with open("data/parts.json", "r") as f:
             part_stats = json.load(f)
         index = player[f"{ship}_parts"].index(oldPart)
         if newPart != "empty":
             if newPart != oldPart:
                 actions -= 1
-                if actions == player['upgrade_apt']-1:
+                if actions == player['upgrade_apt']-1 and discTileUpgrade == "dummy":
                     player_helper.spend_influence_on_action("upgrade")
         else:
             newPart = player[f"old_{ship}_parts"][index]
-        if newPart in ["anm", "axc", "cod", "fls", "hyg", "ins", "iod", "iom", "iot", "jud", "mus", "ricon", "shd", "som", "socha"]:
+        if newPart in ["anm", "axc", "cod", "fls", "hyg", "ins", "iod", "iom", "iot", "jud", "mus", "ricon", "shd", "som", "socha"] and newPart in player_helper.stats["ancient_parts"]:
             player_helper.stats["ancient_parts"].remove(newPart)
         player_helper.stats[f"{ship}_parts"][index] = newPart
         shipCheck = PlayerShip(player_helper.stats, ship)
@@ -94,10 +110,11 @@ class UpgradeButtons:
             ships = ["interceptor","cruiser","dread","starbase"]
             for ship2 in ships:
                 view.add_item(Button(label=ship2.capitalize(), style=discord.ButtonStyle.blurple, custom_id=f"FCID{player['color']}_upgradeShip_{str(actions)}_{ship2}"))
-        view.add_item(Button(label="End Turn", style=discord.ButtonStyle.red, custom_id="endTurn"))
+        view.add_item(Button(label="End Turn", style=discord.ButtonStyle.red, custom_id=f"FCID{player['color']}_endTurn"))
         await interaction.message.delete()
-        await interaction.response.send_message(f"{interaction.user.mention} replaced {part_stats[oldPart]['name']} with {part_stats[newPart]['name']} on their {ship.capitalize()} which now look like this",file=drawing.show_player_ship_area(image))
-        await interaction.followup.send(
-            f"{interaction.user.mention}, choose which ship you would like to upgrade or end turn.", view=view)
+        await interaction.response.send_message(f"{interaction.user.mention} replaced {part_stats[oldPart]['name']} with {part_stats[newPart]['name']} on their {ship.capitalize()} which now looks like this",file=drawing.show_player_ship_area(image))
+        if discTileUpgrade == "dummy":
+            await interaction.followup.send(
+                f"{interaction.user.mention}, choose which ship you would like to upgrade or end turn.", view=view)
